fix: consistency in naming and directories
