+++ src/steps/data_validation/processing/processing_bounding_box_cropper_data_validator.py
@@ -2,12 +2,12 @@ from src import Pipeline
 from src import step
 from src.models.contexts.picsellia_context import PicselliaProcessingContext
 from src.models.dataset.dataset_context import DatasetContext
-from src.models.parameters.processing.bounding_box_cropper_parameters import (
-    BoundingBoxCropperParameters,
+from src.models.parameters.processing.processing_bounding_box_cropper_parameters import (
+    ProcessingBoundingBoxCropperParameters,
 )
 
-from src.models.steps.data_validation.processing.bounding_box_cropper_data_validator import (
-    BoundingBoxCropperDataValidator,
+from src.models.steps.data_validation.processing.processing_bounding_box_cropper_data_validator import (
+    ProcessingBoundingBoxCropperDataValidator,
 )
 
 
@@ -16,10 +16,10 @@ def bounding_box_cropper_data_validator(
     dataset_context: DatasetContext,
 ) -> None:
     context: PicselliaProcessingContext[
-        BoundingBoxCropperParameters
+        ProcessingBoundingBoxCropperParameters
     ] = Pipeline.get_active_context()
 
-    validator = BoundingBoxCropperDataValidator(
+    validator = ProcessingBoundingBoxCropperDataValidator(
         dataset_context=dataset_context,
         client=context.client,
         label_name_to_extract=context.processing_parameters.label_name_to_extract,
